Fix permission issues: View and Update permission does not work across users, User is always 'alice', 'Delete' available even when user does not have permission.
A user cannot view another's permission even though the user has view permissions.
The same problem occurs with delete.  Deletion is done on the browser side but
the store returns message that deletion is not possible.
User is also always alice.
We need to setup Consumer correctly, otherwise, it does not have the correct
key, etc. and the above 3 issues occur.
Finally, when a user has no 'update/delete' permission, the delete option is
still displayed.  We did not check for 'delete' permission which almost
never exists, we should also check for 'update'.